Introduce namespaced/namespacedCfg for consistent option naming
It's annoying having to hardcode the full option structure in each file,
risking typos and just general mistakes as well as having to change them
everywhere when moved.

This moves the responsibility of naming the options to a function that
can be called the same way in each module.

This also means that moving files changes the option paths, so there
needs to be a migration path which is already possible in today's NixOS
module system.